Reduce usage of `this.` qualifier in generated code
Only keep it where it is needed to distinguish from local variable or parameter.

Though on the other hand the `this.` makes it more obvious in the code
generator and in the generated code that it refers to an instance member.
Maybe add it again in the future, but then apply it consistently to all
instance member accesses.